pyproject.toml
[tool.poetry]
name = "studio-ghibli-tracker"
version = "0.1.0"
description = ""
authors = ["None"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.1"
SQLAlchemy = "^1.4.35"
uvicorn = "^0.17.6"
python-dotenv = "^0.20.0"
alembic = "^1.7.7"
psycopg2-binary = "^2.9.3"
sqlmodel = "^0.0.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"